src/requirements.txt
python-dotenv==1.0.1
langchain==0.3.0
langchain_community==0.3.0
langchain_text-splitters==0.3.0
langchain_core==0.3.0
langchain_experimental==0.3.0
langchain_openai==0.2.0
wikipedia==1.4.0
